+++ pkg/node/bypass.go
@@ -6,36 +6,36 @@ import (
 	"github.com/mariomac/pipes/pkg/node/internal/connect"
 )
 
-// Bypass node just makes sure, at graph construction time, that the inputs of this node
+// bypass node just makes sure, at graph construction time, that the inputs of this node
 // are bypassed to the destination nodes.
-// At a logical level, you can see a Bypass node as a Middle[T, T] node that just forwards
+// At a logical level, you can see a bypass node as a middle[T, T] node that just forwards
 // its input to the output channel.
-// At an implementation level, Bypass[T] is much more efficient because it just makes sure
+// At an implementation level, bypass[T] is much more efficient because it just makes sure
 // that its input channel is connected to its destination nodes, without adding any extra
 // goroutine nor channel operation.
-// Bypass is useful for implementing constructors that might return an optional Middle[T, T] node
-// (according to e.g. the user configuration) or just a Bypass[T] node to transparently
+// bypass is useful for implementing constructors that might return an optional middle[T, T] node
+// (according to e.g. the user configuration) or just a bypass[T] node to transparently
 // forward data to the destination nodes.
-type Bypass[INOUT any] struct {
+type bypass[INOUT any] struct {
 	outs []Receiver[INOUT]
 }
 
-func (b *Bypass[INOUT]) SendTo(r ...Receiver[INOUT]) {
+func (b *bypass[INOUT]) SendTo(r ...Receiver[INOUT]) {
 	b.outs = append(b.outs, r...)
 }
 
-func (b *Bypass[INOUT]) InType() reflect.Type {
+func (b *bypass[INOUT]) InType() reflect.Type {
 	return b.OutType()
 }
 
-func (b *Bypass[INOUT]) OutType() reflect.Type {
+func (b *bypass[INOUT]) OutType() reflect.Type {
 	var v INOUT
 	return reflect.TypeOf(v)
 }
 
 // nolint:unused
 // golangci-lint bug: it's actually used through its interface
-func (b *Bypass[INOUT]) isStarted() bool {
+func (b *bypass[INOUT]) isStarted() bool {
 	// returns true if all the forwarded nodes are started
 	started := true
 	for _, o := range b.outs {
@@ -46,9 +46,9 @@ func (b *Bypass[INOUT]) isStarted() bool {
 
 // nolint:unused
 // golangci-lint bug: it's actually used through its interface
-func (b *Bypass[INOUT]) start() {
+func (b *bypass[INOUT]) start() {
 	if len(b.outs) == 0 {
-		panic("Bypass node should have outputs")
+		panic("bypass node should have outputs")
 	}
 	for _, o := range b.outs {
 		if !o.isStarted() {
@@ -59,7 +59,7 @@ func (b *Bypass[INOUT]) start() {
 
 // nolint:unused
 // golangci-lint bug: it's actually used through its interface
-func (b *Bypass[INOUT]) joiners() []*connect.Joiner[INOUT] {
+func (b *bypass[INOUT]) joiners() []*connect.Joiner[INOUT] {
 	joiners := make([]*connect.Joiner[INOUT], 0, len(b.outs))
 	for _, o := range b.outs {
 		joiners = append(joiners, o.joiners()...)