+++ pkg/node/bypass_test.go
@@ -1,28 +1,31 @@
-package node
+package node_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/mariomac/pipes/pkg/node"
 	helpers "github.com/mariomac/pipes/pkg/test"
 )
 
 func TestBypass_Single(t *testing.T) {
-	start := AsStart[int](func(out chan<- int) {
+	// TODO: pipe.New()
+	p := node.NewPipe()
+	start := node.AddStart(p, func(out chan<- int) {
 		out <- 1
 		out <- 2
 		out <- 3
 	})
-	bypass := &Bypass[int]{}
+	bypass := node.AddMiddleOpt[int](p, nil)
 	var recv []int
-	term := AsTerminal[int](func(in <-chan int) {
+	term := node.AddTerminal(p, func(in <-chan int) {
 		for i := range in {
 			recv = append(recv, i)
 		}
 	})
 	var recvMul []int
-	termMul := AsTerminal[int](func(in <-chan int) {
+	termMul := node.AddTerminal(p, func(in <-chan int) {
 		for i := range in {
 			recvMul = append(recvMul, 10*i)
 		}
@@ -30,28 +33,32 @@ func TestBypass_Single(t *testing.T) {
 	start.SendTo(bypass)
 	bypass.SendTo(term, termMul)
 
-	start.Start()
-	helpers.ReadChannel(t, DoneAll(term, termMul), timeout)
+	p.Start()
+	helpers.ReadChannel(t, p.Done(), timeout)
 
 	assert.Equal(t, []int{1, 2, 3}, recv)
 	assert.Equal(t, []int{10, 20, 30}, recvMul)
 }
 
 func TestBypass_Multi(t *testing.T) {
-	start := AsStart[int](func(out chan<- int) {
+	p := node.NewPipe()
+
+	start := node.AddStart(p, func(out chan<- int) {
 		out <- 1
 		out <- 2
 		out <- 3
 	})
-	bypass1, bypass2, bypass3 := &Bypass[int]{}, &Bypass[int]{}, &Bypass[int]{}
+	bypass1 := node.AddMiddleOpt[int](p, nil)
+	bypass2 := node.AddMiddleOpt[int](p, nil)
+	bypass3 := node.AddMiddleOpt[int](p, nil)
 	var recv []int
-	term := AsTerminal[int](func(in <-chan int) {
+	term := node.AddTerminal(p, func(in <-chan int) {
 		for i := range in {
 			recv = append(recv, i)
 		}
 	})
 	var recvMul []int
-	termMul := AsTerminal[int](func(in <-chan int) {
+	termMul := node.AddTerminal(p, func(in <-chan int) {
 		for i := range in {
 			recvMul = append(recvMul, 10*i)
 		}
@@ -61,33 +68,36 @@ func TestBypass_Multi(t *testing.T) {
 	bypass2.SendTo(bypass3)
 	bypass3.SendTo(term, termMul)
 
-	start.Start()
-	helpers.ReadChannel(t, DoneAll(term, termMul), timeout)
+	p.Start()
+	helpers.ReadChannel(t, p.Done(), timeout)
 
 	assert.Equal(t, []int{1, 2, 3}, recv)
 	assert.Equal(t, []int{10, 20, 30}, recvMul)
 }
 
 func TestBypass_Mixed(t *testing.T) {
-	start := AsStart[int](func(out chan<- int) {
+	p := node.NewPipe()
+	start := node.AddStart(p, func(out chan<- int) {
 		out <- 1
 		out <- 2
 		out <- 3
 	})
-	bypass1, bypass2, bypass3 := &Bypass[int]{}, &Bypass[int]{}, &Bypass[int]{}
-	mul := AsMiddle(func(in <-chan int, out chan<- int) {
+	bypass1 := node.AddMiddleOpt[int](p, nil)
+	bypass2 := node.AddMiddleOpt[int](p, nil)
+	bypass3 := node.AddMiddleOpt[int](p, nil)
+	mul := node.AddMiddle(p, func(in <-chan int, out chan<- int) {
 		for i := range in {
 			out <- i * 10
 		}
 	})
 	var recv []int
-	term := AsTerminal[int](func(in <-chan int) {
+	term := node.AddTerminal(p, func(in <-chan int) {
 		for i := range in {
 			recv = append(recv, i)
 		}
 	})
 	var recvAdd []int
-	termAdd := AsTerminal[int](func(in <-chan int) {
+	termAdd := node.AddTerminal(p, func(in <-chan int) {
 		for i := range in {
 			recvAdd = append(recvAdd, 1+i)
 		}
@@ -99,8 +109,8 @@ func TestBypass_Mixed(t *testing.T) {
 	bypass3.SendTo(term)
 	mul.SendTo(termAdd)
 
-	start.Start()
-	helpers.ReadChannel(t, DoneAll(term, termAdd), timeout)
+	p.Start()
+	helpers.ReadChannel(t, p.Done(), timeout)
 
 	assert.Equal(t, []int{1, 2, 3}, recv)
 	assert.Equal(t, []int{11, 21, 31}, recvAdd)